чистовой вариант
