--- src/generate-card.js
@@ -1,384 +1,0 @@
-import {
-  formatHachtags
-} from "./format-hachtags";
-
-import {
-  days
-} from "./mock/repeating-days";
-
-const dateFormatter = new Intl.DateTimeFormat(`en-US`, {
-  day: `numeric`
-});
-const monthFormatter = new Intl.DateTimeFormat(`en-US`, {
-  month: `long`
-});
-
-export const getCardTemplate = (card) => {
-  return `<article class="card card--${card._color} ${card._isRepeating() ? `card--repeat` : ``}" id="${card._id}">
-    <form class="card__form" method="get">
-      <div class="card__inner">
-        <div class="card__control">
-          <button type="button" class="card__btn card__btn--edit">
-            edit
-          </button>
-          <button type="button" class="card__btn card__btn--archive">
-            archive
-          </button>
-          <button
-            type="button"
-            class="card__btn card__btn--favorites card__btn--disabled"
-          >
-            favorites
-          </button>
-        </div>
-
-        <div class="card__color-bar">
-          <svg class="card__color-bar-wave" width="100%" height="10">
-            <use xlink:href="#wave"></use>
-          </svg>
-        </div>
-
-        <div class="card__textarea-wrap">
-          <label>
-            <textarea
-              class="card__text"
-              placeholder="Start typing your text here..."
-              name="text"
-            >${card._title}</textarea>
-          </label>
-        </div>
-
-        <div class="card__settings">
-          <div class="card__details">
-            <div class="card__dates">
-              <button class="card__date-deadline-toggle" type="button">
-                date: <span class="card__date-status">no</span>
-              </button>
-
-              <fieldset class="card__date-deadline">
-                <label class="card__input-deadline-wrap">
-                  <input
-                    class="card__date"
-                    type="text"
-                    placeholder="${dateFormatter.format(card._dueDate)} ${monthFormatter.format(card._dueDate)}"
-                    name="date"
-                    value = "${dateFormatter.format(card._dueDate)} ${monthFormatter.format(card._dueDate)}"
-                  />
-                </label>
-                <label class="card__input-deadline-wrap">
-                  <input
-                    class="card__time"
-                    type="text"
-                    placeholder="${card._dueDate.toLocaleTimeString(`en-US`, {hour: `numeric`, minute: `numeric`})}"
-                    name="time"
-                    value = "${card._dueDate.toLocaleTimeString(`en-US`, {hour: `numeric`, minute: `numeric`})}"
-                  />
-                </label>
-              </fieldset>
-
-              <button class="card__repeat-toggle" type="button">
-                repeat:<span class="card__repeat-status">no</span>
-              </button>
-
-              <fieldset class="card__repeat-days">
-                <div class="card__repeat-days-inner">
-                  <input
-                    class="visually-hidden card__repeat-day-input"
-                    type="checkbox"
-                    id="repeat-mo-5"
-                    name="repeat"
-                    value="mo"
-                    ${card._repeatingDays[days.Monday] ? `checked` : ``}
-                  />
-                  <label class="card__repeat-day" for="repeat-mo-5"
-                    >mo</label
-                  >
-                  <input
-                    class="visually-hidden card__repeat-day-input"
-                    type="checkbox"
-                    id="repeat-tu-5"
-                    name="repeat"
-                    value="tu"
-                    ${card._repeatingDays[days.Tuesday] ? `checked` : ``}
-                  />
-                  <label class="card__repeat-day" for="repeat-tu-5"
-                    >tu</label
-                  >
-                  <input
-                    class="visually-hidden card__repeat-day-input"
-                    type="checkbox"
-                    id="repeat-we-5"
-                    name="repeat"
-                    value="we"
-                    ${card._repeatingDays[days.Wednesday] ? `checked` : ``}
-                  />
-                  <label class="card__repeat-day" for="repeat-we-5"
-                    >we</label
-                  >
-                  <input
-                    class="visually-hidden card__repeat-day-input"
-                    type="checkbox"
-                    id="repeat-th-5"
-                    name="repeat"
-                    value="th"
-                    ${card._repeatingDays[days.Thursday] ? `checked` : ``}
-                  />
-                  <label class="card__repeat-day" for="repeat-th-5"
-                    >th</label
-                  >
-                  <input
-                    class="visually-hidden card__repeat-day-input"
-                    type="checkbox"
-                    id="repeat-fr-5"
-                    name="repeat"
-                    value="fr"
-                    ${card._repeatingDays[days.Friday] ? `checked` : ``}
-                  />
-                  <label class="card__repeat-day" for="repeat-fr-5"
-                    >fr</label
-                  >
-                  <input
-                    class="visually-hidden card__repeat-day-input"
-                    type="checkbox"
-                    name="repeat"
-                    value="sa"
-                    id="repeat-sa-5"
-                    ${card._repeatingDays[days.Saturday] ? `checked` : ``}
-                  />
-                  <label class="card__repeat-day" for="repeat-sa-5"
-                    >sa</label
-                  >
-                  <input
-                    class="visually-hidden card__repeat-day-input"
-                    type="checkbox"
-                    id="repeat-su-5"
-                    name="repeat"
-                    value="su"
-                    ${card._repeatingDays[days.Sunday] ? `checked` : ``}
-                  />
-                  <label class="card__repeat-day" for="repeat-su-5"
-                    >su</label
-                  >
-                </div>
-              </fieldset>
-            </div>
-
-            <div class="card__hashtag">
-              <div class="card__hashtag-list">
-                ${formatHachtags(card._tags)}
-              </div>
-
-              <label>
-                <input
-                  type="text"
-                  class="card__hashtag-input"
-                  name="hashtag-input"
-                  placeholder="Type new hashtag here"
-                />
-              </label>
-            </div>
-          </div>
-
-          <label class="card__img-wrap">
-            <input
-              type="file"
-              class="card__img-input visually-hidden"
-              name="img"
-            />
-            <img
-              src="${card._picture}"
-              alt="task picture"
-              class="card__img"
-            />
-          </label>
-
-          <div class="card__colors-inner">
-            <h3 class="card__colors-title">Color</h3>
-            <div class="card__colors-wrap">
-              <input
-                type="radio"
-                id="color-black-5"
-                class="card__color-input card__color-input--black visually-hidden"
-                name="color"
-                value="black"
-              />
-              <label
-                for="color-black-5"
-                class="card__color card__color--black"
-                >black</label
-              >
-              <input
-                type="radio"
-                id="color-yellow-5"
-                class="card__color-input card__color-input--yellow visually-hidden"
-                name="color"
-                value="yellow"
-              />
-              <label
-                for="color-yellow-5"
-                class="card__color card__color--yellow"
-                >yellow</label
-              >
-              <input
-                type="radio"
-                id="color-blue-5"
-                class="card__color-input card__color-input--blue visually-hidden"
-                name="color"
-                value="blue"
-              />
-              <label
-                for="color-blue-5"
-                class="card__color card__color--blue"
-                >blue</label
-              >
-              <input
-                type="radio"
-                id="color-green-5"
-                class="card__color-input card__color-input--green visually-hidden"
-                name="color"
-                value="green"
-                checked
-              />
-              <label
-                for="color-green-5"
-                class="card__color card__color--green"
-                >green</label
-              >
-              <input
-                type="radio"
-                id="color-pink-5"
-                class="card__color-input card__color-input--pink visually-hidden"
-                name="color"
-                value="pink"
-              />
-              <label
-                for="color-pink-5"
-                class="card__color card__color--pink"
-                >pink</label
-              >
-            </div>
-          </div>
-        </div>
-
-        <div class="card__status-btns">
-          <button class="card__save" type="submit">save</button>
-          <button class="card__delete" type="button">delete</button>
-        </div>
-      </div>
-    </form>
-  </article>`;
-};
-
-export const getEditCardtemplate = (card) => {
-  return `<article class="card card--edit card--${card._color} ${card._isRepeating() ? `card--repeat` : ``}" id="${card._id}">
-    <form class="card__form" method="get">
-      <div class="card__inner">
-        <div class="card__control">
-          <button type="button" class="card__btn card__btn--edit">edit</button>
-          <button type="button" class="card__btn card__btn--archive">archive</button>
-          <button type="button" class="card__btn card__btn--favorites card__btn--disabled">favorites</button>
-        </div>
-
-        <div class="card__color-bar">
-          <svg class="card__color-bar-wave" width="100%" height="10">
-            <use xlink:href="#wave"></use>
-          </svg>
-        </div>
-
-        <div class="card__textarea-wrap">
-          <label>
-            <textarea class="card__text" placeholder="Start typing your text here..." name="text">${card._title}</textarea>
-          </label>
-        </div>
-
-        <div class="card__settings">
-          <div class="card__details">
-            <div class="card__dates">
-              <button class="card__date-deadline-toggle" type="button">
-                date: <span class="card__date-status">no</span>
-              </button>
-
-              <fieldset class="card__date-deadline" disabled>
-                <label class="card__input-deadline-wrap">
-                  <input class="card__date" type="text" placeholder="23 September" name="date" />
-                </label>
-
-                <label class="card__input-deadline-wrap">
-                  <input class="card__time" type="text" placeholder="11:15 PM" name="time" />
-                </label>
-              </fieldset>
-
-              <button class="card__repeat-toggle" type="button">
-                repeat: <span class="card__repeat-status">no</span>
-              </button>
-
-              <fieldset class="card__repeat-days" disabled>
-                <div class="card__repeat-days-inner">
-                  <input class="visually-hidden card__repeat-day-input" type="checkbox" id="repeat-mo-5" name="repeat" value="mo" ${card._repeatingDays[days.Monday] ? `checked` : ``} />
-                  <label class="card__repeat-day" for="repeat-mo-5">mo</label>
-
-                  <input class="visually-hidden card__repeat-day-input" type="checkbox" id="repeat-tu-5" name="repeat" value="tu" ${card._repeatingDays[days.Tuesday] ? `checked` : ``} />
-                  <label class="card__repeat-day" for="repeat-tu-5">tu</label>
-
-                  <input class="visually-hidden card__repeat-day-input" type="checkbox" id="repeat-we-5" name="repeat" value="we" ${card._repeatingDays[days.Wednesday] ? `checked` : ``} />
-                  <label class="card__repeat-day" for="repeat-we-5" >w</label>
-
-                  <input class="visually-hidden card__repeat-day-input" type="checkbox" id="repeat-th-5" name="repeat" value="th" ${card._repeatingDays[days.Thursday] ? `checked` : ``} />
-                  <label class="card__repeat-day" for="repeat-th-5">th</label>
-
-                  <input class="visually-hidden card__repeat-day-input" type="checkbox" id="repeat-fr-5" name="repeat" value="fr" ${card._repeatingDays[days.Friday] ? `checked` : ``} />
-                  <label class="card__repeat-day" for="repeat-fr-5" >fr</label>
-
-                  <input class="visually-hidden card__repeat-day-input" type="checkbox" name="repeat" value="sa" id="repeat-sa-5" ${card._repeatingDays[days.Saturday] ? `checked` : ``} />
-                  <label class="card__repeat-day" for="repeat-sa-5">sa</label>
-
-                  <input class="visually-hidden card__repeat-day-input" type="checkbox" id="repeat-su-5" name="repeat" value="su" ${card._repeatingDays[days.Sunday] ? `checked` : ``} />
-                  <label class="card__repeat-day" for="repeat-su-5" >su</label>
-                </div>
-              </fieldset>
-            </div>
-
-            <div class="card__hashtag">
-              <div class="card__hashtag-list">
-              ${formatHachtags(card._tags)}
-              </div>
-
-              <label>
-                <input type="text" class="card__hashtag-input" name="hashtag-input" placeholder="Type new hashtag here" />
-              </label>
-            </div>
-          </div>
-
-          <label class="card__img-wrap card__img-wrap--empty">
-            <input type="file" class="card__img-input visually-hidden" name="img" />
-          </label>
-
-          <div class="card__colors-inner">
-            <h3 class="card__colors-title">Color</h3>
-            <div class="card__colors-wrap">
-              <input type="radio" id="color-black-5" class="card__color-input card__color-input--black visually-hidden" name="color" value="black" />
-              <label for="color-black-5" class="card__color card__color--black">black</label>
-
-              <input type="radio" id="color-yellow-5" class="card__color-input card__color-input--yellow visually-hidden" name="color" value="yellow" />
-              <label for="color-yellow-5" class="card__color card__color--yellow">yellow</label>
-
-              <input type="radio" id="color-blue-5" class="card__color-input card__color-input--blue visually-hidden" name="color" value="blue" />
-              <label for="color-blue-5" class="card__color card__color--blue">blue</label>
-
-              <input type="radio" id="color-green-5" class="card__color-input card__color-input--green visually-hidden" name="color" value="green" checked />
-              <label for="color-green-5" class="card__color card__color--green">green</label>
-
-              <input type="radio" id="color-pink-5" class="card__color-input card__color-input--pink visually-hidden" name="color" value="pink" />
-              <label for="color-pink-5" class="card__color card__color--pink">pink</label>
-            </div>
-          </div>
-        </div>
-
-        <div class="card__status-btns">
-          <button class="card__save" type="submit">save</button>
-          <button class="card__delete" type="button">delete</button>
-        </div>
-      </div>
-    </form>
-  </article>`;
-};
